+++ playground/src/plugins/index.ts
@@ -1,7 +1,49 @@
+import {
+  multiVariableText,
+  text,
+  barcodes,
+  image,
+  svg,
+  line,
+  table,
+  rectangle,
+  ellipse,
+  dateTime,
+  date,
+  time,
+  select,
+  checkbox,
+  radioGroup,
+} from '@pdfme/schemas';
 import { signature } from './signature';
 
-const plugins = {
-  signature,
+export const getPlugins = () => {
+  return {
+    Text: text,
+    'Multi-Variable Text': multiVariableText,
+    Table: table,
+    Line: line,
+    Rectangle: rectangle,
+    Ellipse: ellipse,
+    Image: image,
+    SVG: svg,
+    Signature: signature,
+    QR: barcodes.qrcode,
+    DateTime: dateTime,
+    Date: date,
+    Time: time,
+    Select: select,
+    Checkbox: checkbox,
+    RadioGroup: radioGroup,
+    // JAPANPOST: barcodes.japanpost,
+    EAN13: barcodes.ean13,
+    // EAN8: barcodes.ean8,
+    // Code39: barcodes.code39,
+    Code128: barcodes.code128,
+    // NW7: barcodes.nw7,
+    // ITF14: barcodes.itf14,
+    // UPCA: barcodes.upca,
+    // UPCE: barcodes.upce,
+    // GS1DataMatrix: barcodes.gs1datamatrix,
+  };
 };
-
-export default plugins;
